Removed delete repo button from NoteForm and moved it into separate component: DangerZone
