+++ src/components/NoteForm.js
@@ -1,5 +1,4 @@
 import React, {useState} from "react";
-import {Redirect} from "react-router";
 import Button from "../styles/Button";
 import {NoteArea, RenderedNote} from "../styles/TextArea";
 import Card from "./Card";
@@ -13,7 +12,6 @@ function NoteForm({goalId, repoName, note}) {
   const [previouslySavedValue, setPreviouslySavedValue] = useState(note);
   const [input, setInput] = useState(note);
   const [editing, setEditing] = useState(false);
-  const [deleted, setDeleted] = useState(false);
 
   const _handleNoteUpdate = () => {
     api
@@ -21,16 +19,6 @@ function NoteForm({goalId, repoName, note}) {
       .then(() => {
         _handleToggleEditing();
         setPreviouslySavedValue(input);
-      })
-      .catch(err => console.log(err));
-  };
-
-  const _handleRepoDeletion = () => {
-    api
-      .updateGoal(goalId, repoName, "CLOSED", input)
-      .then(() => {
-        setEditing(false);
-        setDeleted(true);
       })
       .catch(err => console.log(err));
   };
@@ -48,7 +36,7 @@ function NoteForm({goalId, repoName, note}) {
     setInput(e.target.value);
   };
 
-  return !deleted ? (
+  return (
     <Card>
       {!editing ? (
         <RenderedNote data-testid="notes-content" >
@@ -81,15 +69,9 @@ function NoteForm({goalId, repoName, note}) {
           <Button primary onClick={_handleCancelEditing}>
             Cancel
           </Button>
-        ) : (
-          <Button primary onClick={_handleRepoDeletion}>
-            Delete
-          </Button>
-        )}
+        ) : ""}
       </FlexCenter>
     </Card>
-  ) : (
-    <Redirect to="/" />
   );
 }
